pyproject.toml
[project]
name = "commercial-vpn-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "passlib[argon2,bcrypt]>=1.7",
  "httpx>=0.27",
  "email-validator>=2.3.0",
  "python-multipart>=0.0.21",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
